requirements.txt
aioflask==0.4.0
anyio==4.0.0
asgiref==3.7.2
blinker==1.6.3
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
greenlet==3.0.1
greenletio==0.10.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-telegram-bot==20.6
sniffio==1.3.0
typing_extensions==4.8.0
uvicorn==0.23.2
Werkzeug==3.0.1
